# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "zod-openapi-play",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice@example.com"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Alice"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice@example.com"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Alice"
          }
        },
        "required": [
          "email",
          "name"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/users/{id}": {
      "get": {
        "summary": "Get a user by id",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create a user",
        "tags": [
          "users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  }
}